% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dataSummary}
\alias{dataSummary}
\title{Run a simple summary on the data}
\usage{
dataSummary(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A data.frame
}
\description{
Run a simple summary on the data
}
