% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getData}
\alias{getData}
\title{Get some data from a public PostgreSQL instance}
\usage{
getData()
}
\value{
A data.frame with the first 100 entries
}
\description{
This is really only to demonstrate that we can read from
PostgreSQL without having to set up our own server
}
