\name{lim.tiny-package}
\alias{lim.tiny-package}
\alias{lim.tiny}
\docType{package}
\title{\packageTitle{lim.tiny}}
\description{\packageDescription{lim.tiny}}
\section{Package Content}{\packageIndices{lim.tiny}}
\author{\packageAuthor{lim.tiny}}
\section{Maintainer}{\packageMaintainer{lim.tiny}}
\keyword{package}
